requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
rank_bm25
llama-cpp-python
numpy
pydantic
